rag-platform/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy>=2.0.40,<3
psycopg[binary]>=3.2.10,<4
alembic==1.14.0
pydantic>=2.11,<3
pydantic-settings>=2.7,<3
python-multipart==0.0.17
redis==5.2.1
arq==0.26.1
httpx==0.28.1
qdrant-client>=1.14,<2
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langchain-community==0.3.13
pypdf==5.1.0
python-docx==1.1.2
charset-normalizer==3.4.0
rank-bm25==0.2.2
